requirements.txt
aiohttp==3.9.5
python-telegram-bot==20.7
oandapyV20==0.6.3
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
transformers==4.41.1
torch==2.3.0
tqdm==4.66.4
numpy==1.26.4
scikit-learn==1.5.0